CV/Unet_Segmentation_2/requirements.txt
numpy
pandas
matplotlib
seaborn
plotly
requests
tqdm
pillow
gradio
opencv-python
onnx
torchsummary
segmentation_models_pytorch
